Create Date (Hotel nights calculator)
This code was a part of a bigger project with some of my acquaintances in my college. The main code was a hotel system which helps clients book rooms from the system directly. This piece of code specifically takes the arrival date and the departer date from the client to calculate the number of nights they're staying. The number of nights is then used to calculate the price of the client's stay along with other factors (room luxury level).